requirements.txt
#celery==4.4.1
#redis==3.4.1
Django==3.2.9
django-ckeditor==6.3.2
django-js-asset==2.0.0
django-modeltranslation==0.17.5
Pillow==9.1.0
psycopg2>=2.7.5,<2.8.0
celery==4.4.1
redis==3.4.1